+++ next.config.js
@@ -37,6 +37,12 @@ const nextConfig = {
       },
       {
         protocol: "https",
+        hostname: "lh3.googleusercontent.com",
+        port: "",
+        pathname: "/**",
+      },
+      {
+        protocol: "https",
         hostname: "cloudflare-ipfs.com",
         port: "",
         pathname: "/**"