+++ prisma/migrations/20240306115029_add_additional_role_for_user/migration.sql
@@ -1,0 +1,8 @@
+/*
+  Warnings:
+
+  - Added the required column `ownerId` to the `Organization` table without a default value. This is not possible if the table is not empty.
+
+*/
+-- AlterEnum
+ALTER TYPE "UserRole" ADD VALUE 'organization';