+++ src/components/premise/index.ts
@@ -1,0 +1,1 @@
+export { PremiseListing } from "./PremiseListing";