+++ src/lib/actions/auth/index.ts
@@ -1,2 +1,3 @@
+export * from "./org-register.ts";
 export * from "./reset.ts";
 export * from "./signup.ts";