+++ src/lib/utils/format-language.ts
@@ -1,0 +1,5 @@
+import type { Locale } from "@/navigation";
+
+export function languageFormatter(locale: Locale) {
+  return new Intl.DisplayNames(locale, { type: "language" });
+}