+++ src/lib/validations/auth/index.ts
@@ -1,3 +1,4 @@
+export * from "./org-register.ts";
 export * from "./reset.ts";
 export * from "./signin.ts";
 export * from "./signup.ts";